LayoutTests/fast/selectors/querySelector-window-inactive.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<a>some text</a>
</body>
<script>
description('Test window-inactive in querySelectorAll.');
shouldBe('document.querySelectorAll("a:window-inactive").length', '0');
if (window.testRunner) {
    testRunner.setWindowIsKey(false);
    shouldBe('document.querySelectorAll("a:window-inactive").length', '1');
}
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
