LayoutTests/fast/selectors/querySelector-window-inactive-expected.txt

Test window-inactive in querySelectorAll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("a:window-inactive").length is 0
PASS document.querySelectorAll("a:window-inactive").length is 1
PASS successfullyParsed is true

TEST COMPLETE
some text
